[CLI] Move library function encoding commands to their own folder (#361)
